Wait for help command

+++ Kernel/process.c
@@ -28,7 +28,7 @@ void init_process() {
     process_table[0].occupied = 1;
 }
 
-int start_proc(char *proc_name, void (*function)(int argc, char *argv[])) {
+uint64_t start_proc(char *proc_name, void (*function)(int argc, char *argv[])) {
     int index_proc = get_new_index();
 
     proc process = &process_table[index_proc];
